[Update] the controllers and services.

+++ src/main/java/com/Prograd/EmployeeManagement/Service/OrganisationServiceImpl.java
@@ -1,10 +1,15 @@
 package com.Prograd.EmployeeManagement.Service;
 
+import com.Prograd.EmployeeManagement.Exceptions.NoAccessException;
 import com.Prograd.EmployeeManagement.Exceptions.OrganisationNotFound;
 import com.Prograd.EmployeeManagement.Modals.Organisation;
+import com.Prograd.EmployeeManagement.Repository.AssetRepository;
+import com.Prograd.EmployeeManagement.Repository.EmployeeRepository;
 import com.Prograd.EmployeeManagement.Repository.OrganisationRepository;
+import com.Prograd.EmployeeManagement.Security.EmployeeSecurity;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,17 +18,21 @@ import java.util.List;
 public class OrganisationServiceImpl implements OrganisationService {
     private OrganisationRepository organisationRepository;
 
-    @Autowired
-    private PasswordEncoder passwordEncoder;
-
     public OrganisationServiceImpl(OrganisationRepository organisationRepository) {
         this.organisationRepository = organisationRepository;
     }
 
+    @Autowired
+    private EmployeeRepository employeeRepository;
+
+    @Autowired
+    private AssetRepository assetRepository;
+
+    @Autowired
+    private EmployeeSecurity employeeSecurity;
+
     @Override
     public Organisation saveOrganisation(Organisation organisation) {
-        String encodedPassword = passwordEncoder.encode(organisation.getPassword());
-        organisation.setPassword(encodedPassword);
         return organisationRepository.save(organisation);
     }
 
@@ -39,19 +48,33 @@ public class OrganisationServiceImpl implements OrganisationService {
     }
 
     @Override
-    public Organisation updateOrganisation(Organisation organisation, int id) throws OrganisationNotFound {
+    public Organisation updateOrganisation(Organisation organisation, int id) throws OrganisationNotFound, NoAccessException {
         Organisation existingOrganisation = organisationRepository.findById(id).orElseThrow(()->new OrganisationNotFound(("Organisation not exist")));
-
         existingOrganisation.setOrganisation_name(organisation.getOrganisation_name());
-        existingOrganisation.setUsername(organisation.getUsername());
-        existingOrganisation.setPassword(organisation.getPassword());
 
         return organisationRepository.save(existingOrganisation);
     }
 
     @Override
-    public void deleteOrganisation(int id) {
-        organisationRepository.deleteById(id);
+    public void deleteOrganisation(int id) throws OrganisationNotFound, NoAccessException {
+        Organisation organisation = organisationRepository.findById(id).orElseThrow(() -> new OrganisationNotFound("Organisation not exist"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        boolean hasEmpRole = authentication.getAuthorities().stream().anyMatch(r -> r.getAuthority().equals("ROLE_EMPLOYEE"));
+
+        if(!hasEmpRole) {
+            boolean res = employeeSecurity.hasOrgId(id);
+            if (res) {
+                employeeRepository.deleteAllEmpByOrgId(organisation);
+                assetRepository.deleteAllAssetByOrgId(organisation);
+                organisationRepository.deleteById(id);
+            } else {
+                throw new NoAccessException("You have no access for this request");
+            }
+        }else{
+            throw new NoAccessException("You have no access for this request");
+        }
     }
 }
